refactor(create-mc-app): restructure package scripts
